+++ NEWS.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* the Segre manifold
+* the warped metric
 * The manifold `HeisenbergMatrices` as the underlying manifold of `HeisenbergGroup`.
 
 ### Changed